tests: compare digests in test_peer_recovery
Problem

Test fails when comparing the first WAL segment because
the system id in the segment header is different. The system id
is not consistently set correctly since segments are usually inited
on the safekeeper sync step with sysid 0.

Summary of Chnages

Compare timeline digests instead. This skips the header.