compute: Disconnect if no response to a pageserver request is received
We've seen some cases in production where a compute doensn't get a
response to a pageserver request for several minutes, or even more. We
haven't found the root cause for that yet, but whatever the reason is,
it seems overly optimistic to think that if the pageserver hasn't
responded for 30 s, we'd get a response if we just wait patiently a
little longer. More likely, the pageserver is dead or there's some
kind of a network glitch so that the TCP connection is dead, or at
least stuck for a long time. Either way, it's better to disconnect and
reconnect. I set the default timeout to 30 s, which should be enough
for any GetPage request under normal circumstances, even if the
pageserver has to download several layer files from remote storage.

Make the disconnect timeout configurable. Also make the "log
interval", after which we print a message to the log configurable, so
that if you change the disconnect timeout, you can set the log timeout
correspondingly. The default log interval is still 10 s. The new GUCs
are called "neon.pageserver_response_log_timeout" and
"neon.pageserver_response_disconnect_timeout".

Includes a basic test for the log and disconnect timeouts.

Implements issue #10857