Split plv8 build into two parts
Plv8 consists of two parts:
1. the V8 engine, which is built from vendored sources, and
2. the PostgreSQL extension.

Split those into two separate steps in the Dockerfile. The first step
doesn't need any PostgreSQL sources or any other files from the neon
repository, just the build tools and the upstream plv8 sources. Use
the build-deps image as the base for that step, so that the layer can
be cached and doesn't need to be rebuilt every time. This is
worthwhile because the V8 build takes a very long time.