Fix the timeout calculations
It would round down the timeout, which means that the call would wake
up just *before* enough time had passed. That lead to busy-waiting for
up to 1 ms before reaching the timeout. To fix, round up.

This doesn't explain the CI failures, but I noticed this when I added
more verbose logging. The more verbose logging is also included to
help debug the CI failure.