Fix usage of WaitEventSetWait() with timeout
WaitEventSetWait() returns the number of "events" that happened, and
only that many events in the WaitEvent array are updated. When the
timeout is reached, it returns 0 and does not modify the WaitEvent
array at all. We were reading 'event.events' without checking the
return value, which would be uninitialized when the timeout was hit.

No test included, as this is harmless at the moment. But this caused
the test I'm including in PR #10882 to fail. That PR changes the logic
to loop back to retry the PQgetCopyData() call if WL_SOCKET_READABLE
was set. Currently, making an extra call to PQconsumeInput() is
harmless, but with that change in logic, it turns into a busy-wait.