compute_ctl: Add OTEL tracing to incoming HTTP requests and startup
We lost this with the switch to axum for the HTTP server. Add it back.

In addition to just resurrecting the functionality we had before, pass
the tracing context of the /configure HTTP request to the
start_postgres operation that runs in the main thread. This way, the
'start_postgres' and all its sub-spans like getting the basebackup
become children of the HTTP request span. This allows end-to-end
tracing of a compute start, all the way from the proxy to the SQL
queries executed by compute_ctl as part of compute startup.