compute_ctl: Add instrumentation to resizing swap and setting disk quota
In local dev environment, these steps take around 100 ms, and they are
in the critical path of a compute startup on a compute pool hit. I
don't know if it's like that in production, but as first step, add
tracing spans to the functions so that they can be measured more
easily.