store Arc<TimelineMetrics> instead of Arc<Timeline> to avoid risk of reference cycles
Storing Arc<Timeline> inside RequestContext risks reference cycles if
the RequestContext gets stored in an object that is owned directly or
indirectly by Timeline.

So, we wrap the TimelineMetrics into an Arc and propagate that instead.
This makes it easy for future metrics to be access through
RequestContext, like we do for storage_io_size here.

To make the page_service case happy, where we may be dispatching to
a different shard every successive request, and don't want to be cloning
from the shared Timeline::metrics on each request, we pre-clone as
part of the handle cache miss.