pageserver: mark unarchival heatmap layers cold
On unarchival, we update the previous heatmap with all visible layers.
When the primary generates a new heatmap it includes all those layers,
so the secondary will download them. Since they're not actually resident
on the primary (we didn't call the warm up API), they'll never be evicted,
so they remain in the heatmap.

We want these layers in the heatmap, since we might wish to warm-up an
unarchived timeline after a shard migration. However, we don't want them
to be downloaded on the secondary until we've warmed up the primary.

Include these layers in the heatmap and mark them as cold. All heatmap
operations act on non-cold layers apart from the attached location warming up API,
which will download the cold layers. Once the cold layers are downloaded on the primary,
they'll be included in the next heatmap as hot and the secondary starts fetching them too.