utils: optionally enable otel tracing in common logging utils
This patch augments `utils::logging::init` with the ability to set up
OTEL tracing infrastructure. The end goal is for the pageserver to use
this in order to export perf traces.

Note that an entirely different tracing subscriber is used. This is to
avoid interference with the existing tracing set-up.

For now, no service uses this functionality.